+++ public/js/datatables.js
@@ -4,16 +4,6 @@ $(() => {
     
     $.fn.dataTable.moment('MM/DD/YYYY HH:mm');
 
-    $('#openTickets').DataTable({
-        "order": [[0, "asc"]],
-        "columnDefs": [{ targets:1, type: 'natural'}]
-    });
-
-    $('#closedTickets').DataTable({
-        "order": [[0, "desc"]],
-        "columnDefs": [{ targets:1, type: 'natural'}]
-    });
-
     $('#myTickets').DataTable({
         "order": [[4, "asc"]],
         "columnDefs": [{ targets:1, type: 'natural'}]